Add general language support
Use prefixes of language tags (instead of the whole tags) so that tags
like en, en-GB and en_US can be treated in a uniform manner.

Also, do not exclude voices whose language tags do not have the
predefined prefixes; they will be placed in the bottom of the language
list, with their raw language tags being used as display names.